batch limit part: add some limit logic

+++ iqshell/common/limit/limit.go
@@ -3,6 +3,6 @@ package limit
 import "github.com/qiniu/qshell/v2/iqshell/common/data"
 
 type Limit interface {
-	Acquire(count int64) *data.CodeError
-	Release(count int64)
+	Acquire(count int) *data.CodeError
+	Release(count int)
 }
